making secondary srs measurement periodic

+++ openair1/PHY/NR_ESTIMATION/nr_ul_estimation.h
@@ -69,9 +69,9 @@ int nr_est_timing_advance_srs(const NR_DL_FRAME_PARMS *frame_parms,
 int nr_est_toa_ns_srs(NR_DL_FRAME_PARMS *frame_parms,
 		          uint8_t N_arx,
 		          uint8_t N_ap,
-              uint8_t N_symb_srs,
+                          uint8_t N_symb_srs,
 			  int32_t srs_estimated_channel_freq[N_arx][N_ap][frame_parms->ofdm_symbol_size * N_symb_srs],
-			  int32_t *srs_toa_ns);
+			  int16_t *srs_toa_ns);
 
 void nr_pusch_ptrs_processing(PHY_VARS_gNB *gNB,
                               NR_DL_FRAME_PARMS *frame_parms,
